feat: add number visibility toggle feature
Implements a UI feature to hide/show sensitive financial numbers across the application.
- Numbers show first 2 digits followed by asterisks when hidden
- Toggle button in topbar allows users to switch between hidden/visible states
- Applied to Dashboard, Assets, and Reports pages
- Uses React Context for global state management
- No data security implications - purely a visual/UI feature
