+++ apps/web/app/(shell)/TopbarWithUser.tsx
@@ -3,9 +3,11 @@
 import { useMemo } from "react";
 import { Topbar } from "@financelab/ui";
 import { useAuth } from "../../lib/auth-context";
+import { useNumberVisibility } from "../../lib/number-visibility-context";
 
 export default function TopbarWithUser() {
   const { user } = useAuth();
+  const { isVisible, toggleVisibility } = useNumberVisibility();
 
   const userLabel = useMemo(() => {
     if (!user) {
@@ -15,5 +17,11 @@ export default function TopbarWithUser() {
     return label ? `Signed in as ${label}` : "Signed in";
   }, [user]);
 
-  return <Topbar userLabel={userLabel} />;
+  return (
+    <Topbar
+      userLabel={userLabel}
+      onToggleNumberVisibility={toggleVisibility}
+      numbersVisible={isVisible}
+    />
+  );
 }
